+++ src/tensor_ops/matmul/mod.rs
@@ -1,5 +1,8 @@
 #![allow(clippy::type_complexity)]
 
+#[cfg(feature = "f16")]
+mod f16;
+
 pub(super) mod cpu_kernel;
 
 #[cfg(feature = "cuda")]
